# api/swagger.yaml
basePath: /
definitions:
...
  caspercloud_internal_repository.Instance:
    properties:
      cloud_init_data:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_id:
        type: string
      name:
        type: string
      project_id:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
  caspercloud_internal_repository.Project:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
  caspercloud_internal_repository.Task:
    properties:
      created_at:
        type: string
      error:
        type: string
      id:
        type: string
      instance_id:
        type: string
      project_id:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  caspercloud_internal_service.CreateInstanceResult:
    properties:
      instance:
        $ref: '#/definitions/caspercloud_internal_repository.Instance'
      task:
        $ref: '#/definitions/caspercloud_internal_repository.Task'
    type: object
  internal_httpapi.createInstanceRequest:
    properties:
      hostname:
        maxLength: 253
        minLength: 1
        type: string
      image_id:
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
      packages:
        items:
          type: string
        maxItems: 64
        type: array
      run_commands:
        items:
          type: string
        maxItems: 32
        type: array
      ssh_public_key:
        maxLength: 8192
        minLength: 1
        type: string
      username:
        maxLength: 32
        minLength: 1
        type: string
    required:
    - hostname
    - image_id
    - name
    - ssh_public_key
    - username
    type: object
  internal_httpapi.createProjectRequest:
    properties:
      name:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  internal_httpapi.docCreateInstanceAccepted:
    properties:
      data:
        $ref: '#/definitions/caspercloud_internal_service.CreateInstanceResult'
    type: object
...
  internal_httpapi.docInstanceData:
    properties:
      data:
        $ref: '#/definitions/caspercloud_internal_repository.Instance'
    type: object
  internal_httpapi.docInstancesData:
    properties:
      data:
        items:
          $ref: '#/definitions/caspercloud_internal_repository.Instance'
        type: array
    type: object
  internal_httpapi.docProjectData:
    properties:
      data:
        $ref: '#/definitions/caspercloud_internal_repository.Project'
    type: object
  internal_httpapi.docProjectsData:
    properties:
      data:
        items:
          $ref: '#/definitions/caspercloud_internal_repository.Project'
        type: array
    type: object
  internal_httpapi.docStatusOK:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'HTTP API for CasperCloud (multi-tenant projects, images, instances).
    Authenticated routes expect `Authorization: Bearer <JWT>`. Project-scoped routes
    require a token whose `active_project_id` matches the path `projectID` (use login
    with `project_id` or POST /v1/auth/switch-project).'
  title: CasperCloud API
  version: "1.0"
paths:
...
  /v1/projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docProjectsData'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project and adds the caller as owner.
      parameters:
      - description: Project name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.docProjectData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
...
  /v1/projects/{projectID}/instances:
    get:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docInstancesData'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List instances
      tags:
      - instances
    post:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance spec
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.createInstanceRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_httpapi.docCreateInstanceAccepted'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create instance
      tags:
      - instances
  /v1/projects/{projectID}/instances/{instanceID}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance UUID
        in: path
        name: instanceID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete instance
      tags:
      - instances
    get:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance UUID
        in: path
        name: instanceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docInstanceData'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get instance
      tags:
      - instances
  /v1/projects/{projectID}/instances/{instanceID}/reboot:
    post:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance UUID
        in: path
        name: instanceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docStatusOK'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Instance lifecycle action
      tags:
      - instances
  /v1/projects/{projectID}/instances/{instanceID}/start:
    post:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance UUID
        in: path
        name: instanceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docStatusOK'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Instance lifecycle action
      tags:
      - instances
  /v1/projects/{projectID}/instances/{instanceID}/stop:
    post:
      parameters:
      - description: Project UUID
        in: path
        name: projectID
        required: true
        type: string
      - description: Instance UUID
        in: path
        name: instanceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.docStatusOK'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Instance lifecycle action
      tags:
      - instances
securityDefinitions:
  BearerAuth:
    description: JWT from POST /v1/auth/login or /v1/auth/register. Prefix value with
      "Bearer " (e.g. `Bearer eyJhbG...`).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"